+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Show/Hideボタンを分かりやすく変更。
 - Qiitadon用にcode表示調整。
+- Qiitadonのstreaming対応。
 
 ## v1.0.0 (2017-05-25)
 - 1.0.0